+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Activitus Bar Change Log
 
+- Add counter to source control view button
+
 ## v0.0.41 - 2021-03-10
 
 - Improve key bindings support